fix: fix crash when installing with yarn
The package phantomjs is actually just an older version of
phantomjs-prebuilt. This older version has a bug which makes it crash
when trying to install with yarn.

See: https://github.com/yarnpkg/yarn/issues/1016#issuecomment-299230894

In fact, since karma-phantomjs-launcher already installs
phantomjs-prebuilt (and we don't actually use PhantomJS other than
through that module), we don't also need to depend on it at all.